+++ html/index_output.html
@@ -1,0 +1,73 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <title>
+        Testing...
+        </title>
+    </head>
+    <body>
+        <div>
+            <p> Date of Past Workouts </p>
+            <ul>
+                
+                <li> 2/1/2019 8:54:57 </li>
+                
+                <li> 2/5/2019 21:12:48 </li>
+                
+                <li> 2/12/2019 21:17:35 </li>
+                
+                <li> 2/13/2019 18:02:47 </li>
+                
+                <li> 2/17/2019 19:43:24 </li>
+                
+                <li> 2/17/2019 19:59:15 </li>
+                
+                <li> 2/19/2019 21:33:46 </li>
+                
+                <li> 2/22/2019 10:50:19 </li>
+                
+                <li> 2/22/2019 11:20:46 </li>
+                
+                <li> 2/23/2019 21:55:09 </li>
+                
+                <li> 2/23/2019 21:55:47 </li>
+                
+                <li> 2/23/2019 21:57:53 </li>
+                
+                <li> 2/24/2019 16:59:02 </li>
+                
+                <li> 2/24/2019 18:07:37 </li>
+                
+                <li> 2/26/2019 15:02:45 </li>
+                
+                <li> 2/26/2019 17:51:32 </li>
+                
+                <li> 2/26/2019 17:52:03 </li>
+                
+                <li> 2/27/2019 13:55:47 </li>
+                
+                <li> 3/2/2019 10:39:40 </li>
+                
+                <li> 3/2/2019 17:55:15 </li>
+                
+                <li> 3/4/2019 22:16:13 </li>
+                
+                <li> 3/5/2019 8:11:47 </li>
+                
+                <li> 2019-03-07 </li>
+                
+                <li> 3/8/2019 9:47:29 </li>
+                
+                <li> 3/10/2019 12:38:11 </li>
+                
+                <li> 3/10/2019 12:40:08 </li>
+                
+                <li> 3/10/2019 12:52:25 </li>
+                
+                <li> 3/11/2019 8:17:20 </li>
+                
+            </ul>
+        </div>
+    </body>
+
+</html>
